requirements.txt
pyserial==3.3 
minimalmodbus==2.1.1